chore(api): move onAuth and onDisconnect logic to framework code
