+++ packages/create-botonic-app/dev-template/api/src/websocket/index.js
@@ -1,8 +1,11 @@
 import { websocketServerFactory } from '@botonic/api/src/websocket'
+import { dataProviderFactory } from '@botonic/core/lib/esm/data-provider'
 
 import { onAuth } from './onauth'
 import { onConnect } from './onconnect'
 import { onDisconnect } from './ondisconnect'
+
+const dataProvider = dataProviderFactory(process.env.DATA_PROVIDER_URL)
 
 export default websocketServerFactory({
   // eslint-disable-next-line no-undef
@@ -10,4 +13,5 @@ export default websocketServerFactory({
   onConnect,
   onAuth,
   onDisconnect,
+  dataProvider,
 })
